Introduce StreamCards class
First step for streamer and viewer separation.
This class generates the cards to be displayed in the streamer,
and handles the annotation events.

Stream no longer relies on the threading service.
The messageContainers from jwz are still used.
They still share the same template.

Replies in the stream are displayed as root level annotation
cards.

Fix #1916